Aula/esempio3/requirements.txt
flask
google-cloud-firestore
flask-login